delete user, no longer support owner

+++ modules/user_del/user_del.php
@@ -3,11 +3,16 @@
 	if ($id !== null) {
 		$user = new user();
 		$user->setId($id);
+		$user_data = $user->returnOneUser();
 
-		if ($user->delete()) {
-			print  $language["actions"]["success"];
+		if ($user_data["rank"] !== "owner") {
+			if ($user->delete()) {
+				print  $language["actions"]["success"];
+			} else {
+				print  $language["actions"]["failure"];
+			}
 		} else {
-			print  $language["actions"]["failure"];
+			print $language["mod_user"]["owner-delete-failure"];
 		}
 	}else{
 		print  $language["actions"]["error"];
